Tools for Working with Parallel Random Seeds

Tools for Working with Parallel Random Seeds

get_random_seed()

set_random_seed(seed)

next_random_seed(seed = get_random_seed())

is_valid_random_seed(seed)

is_lecyer_cmrg_seed(seed)

as_lecyer_cmrg_seed(seed)

Arguments

seed

A random seed

Value

get_random_seed() returns the current

.Random.seed. If it does not exists, it returns NULL.

set_random_seed(seed) sets a new value on .Random.seed, and invisibly returns the old seed. If seed = NULL, then the .Random.seed is removed.

next_random_seed() updates .Random.seed by drawning an dummy random number internally, and invisibly returns the new seed.

is_valid_random_seed(seed) returns TRUE if seed is a valid random seed of any RNG kind, otherwise FALSE. This function does not update .Random.seed.

is_lecyer_cmrg_seed(seed) returns TRUE if seed is a valid random seed of kind L'Ecuyer-CMRG, otherwise FALSE. This function does not update .Random.seed.

as_lecyer_cmrg_seed(seed) returns L'Ecuyer-CMRG random seed based on random seed seed. If seed is already of the right RNG kind, then that seed is returned as-is. If a scalar, then a L'Ecuyer-CMRG random seed is generated from that seed with the help of set.seed(). This function does not update .Random.seed.

See also

Examples

## This example gives identical results in the two
## approaches only if .Random.seed already exists
seed_org <- next_random_seed()
set.seed(42)

## Draw two random number from 1:100 one after the other
seed_before <- get_random_seed()
x <- sample.int(100L, size = 1L)
y <- sample.int(100L, size = 1L)
seed_after <- get_random_seed()
print(c(x, y))
#> [1] 2 4

## Draw two random number from 1:100 at once
set_random_seed(seed_before)
z <- sample.int(100L, size = 2L)
seed_after2 <- get_random_seed()
print(z)
#> [1] 2 4

## The RNG state is the same after the second ...
## run as after the first, two-step, approach
stopifnot(identical(seed_after2, seed_after))

## .... and the exact same set of random numbers where
## drawn [only true in R (>= 4.0.0)]
if (getRversion() >= "4.0.0") {
  stopifnot(identical(z, c(x, y)))
} else {
  stopifnot(identical(z[seq_along(x)], x))
}

set_random_seed(seed_org)
stopifnot(identical(get_random_seed(), seed_org))