Produce Reproducible Seeds for Parallel Random Number Generation

Produce Reproducible Seeds for Parallel Random Number Generation

make_rng_seeds(count, seed = FALSE)

Arguments

count

The number of RNG seeds to produce.

seed

A logical specifying whether RNG seeds should be generated or not. (seed = NULL corresponds to seed = FALSE). If a list, then it should be of length count and each element should consist of a valid RNG seed.

Value

Returns a non-named list of count independent L'Ecuyer-CMRG

random seeds. If seed is NULL or FALSE, then NULL is returned.

Details

This function generates count independent L'Ecuyer-CMRG random seeds that can be used as .Random.seed for parallel processing. These seeds are produced with help of parallel::nextRNGSubStream() and parallel::nextRNGStream() using a strategy that

seed <- <initial RNG seed>
for (ii in seq_len(count)) {
  seeds[[ii]] <- parallel::nextRNGSubStream(seed)
  seed <- parallel::nextRNGStream(seed)
}

This function forwards the RNG state 1 + count times if seed = TRUE.

Examples

## Set up L'Ecuyer-CMRG random seeds for 100 parallel tasks
## based on .Random.seed in the current R process
seeds <- make_rng_seeds(100L, seed = TRUE)
str(seeds)
#> List of 100
#>  $ : int [1:7] 10407 -343209703 2030968786 754822341 752900433 -634994 -1343441506
#>  $ : int [1:7] 10407 142959816 -296877942 918696220 -1754600167 -1204656886 -1831218389
#>  $ : int [1:7] 10407 827622561 -577885192 731320672 -768022873 338858200 1979967683
#>  $ : int [1:7] 10407 1900137755 -599320842 -793432992 -1594705310 1151792082 413129429
#>  $ : int [1:7] 10407 1574648943 963351853 1411740630 -382608763 1363546495 -1070082930
#>  $ : int [1:7] 10407 176394965 1373791967 132576938 764198855 -2128241825 892015376
#>  $ : int [1:7] 10407 1065399593 -1408671405 1219570132 2090217104 -1833089452 -679972188
#>  $ : int [1:7] 10407 313938871 -1158139092 -1472071674 660653992 -619779358 2046864379
#>  $ : int [1:7] 10407 275448961 1222636144 363802026 220091907 211446504 -383710013
#>  $ : int [1:7] 10407 1692956459 564935216 475698151 576970587 -1949536686 -1803878020
#>  $ : int [1:7] 10407 648214283 1997773988 -33958713 1077559683 -1234274685 -1115413906
#>  $ : int [1:7] 10407 -963359189 549930986 -747254219 423715603 1993773396 -168560936
#>  $ : int [1:7] 10407 -1919035660 1257243759 1465184973 -805526912 -2051466547 86935996
#>  $ : int [1:7] 10407 -924514972 261526697 -2141817114 460804849 -1931122150 988147099
#>  $ : int [1:7] 10407 -405029492 -628113056 55138083 -293578086 -1393693565 350122050
#>  $ : int [1:7] 10407 -1605406824 -517244652 -1087138718 -1469528894 -1070346309 401148167
#>  $ : int [1:7] 10407 -1863803497 -1858556122 832320599 -1494145475 1937630943 1096839058
#>  $ : int [1:7] 10407 2058436461 1986752982 -1983499265 1493072770 -504231243 1138871141
#>  $ : int [1:7] 10407 1961330017 2096091977 -34162751 1213694860 -433383497 1327210636
#>  $ : int [1:7] 10407 372248645 1565689800 577882827 -2144741977 1433870571 174352437
#>  $ : int [1:7] 10407 -1556424718 -563653988 -1151660816 332995724 484865793 -1300070471
#>  $ : int [1:7] 10407 -1375174595 -1727297380 -1114974803 1788886982 2087160342 -1895751282
#>  $ : int [1:7] 10407 559483322 1758506122 -617205032 -2025852000 170773193 629675538
#>  $ : int [1:7] 10407 -1933112140 -1930681567 397432375 -1283118277 1983015549 -1854980789
#>  $ : int [1:7] 10407 663468002 34271196 498141091 739481317 -413160281 1836980280
#>  $ : int [1:7] 10407 813519435 812572360 1112422565 -973774036 -872766810 1804594225
#>  $ : int [1:7] 10407 -1134386881 -1519516596 -1947446741 19945115 -398122005 240314078
#>  $ : int [1:7] 10407 1935068104 -1324840317 308474709 93389455 1455832361 -1993967922
#>  $ : int [1:7] 10407 -1750014759 -1253294776 953491023 1962459642 1020508467 -1252008373
#>  $ : int [1:7] 10407 243399105 -1714220124 1885294993 203285232 -1866136377 -424968074
#>  $ : int [1:7] 10407 -125382000 -353723879 795081581 -94045042 556278995 -1067123202
#>  $ : int [1:7] 10407 -1084557893 555455675 -888530002 -1655713500 2046558449 885554489
#>  $ : int [1:7] 10407 -226012899 1932423230 -914086590 -1485889415 -2067806797 694448560
#>  $ : int [1:7] 10407 -519019983 2146634821 -857910833 -1633955382 -1510114968 734059045
#>  $ : int [1:7] 10407 1286019106 1598370924 2107065340 -642265046 -472438170 383040709
#>  $ : int [1:7] 10407 1072476205 -1927117797 -192149649 -722770391 -763130231 1195973589
#>  $ : int [1:7] 10407 -1034315310 1389374623 -1022968793 1249413982 -986379592 -1955334808
#>  $ : int [1:7] 10407 1035307755 -1137672606 1664002036 -2141810856 285284385 -1358666514
#>  $ : int [1:7] 10407 -1778288166 -1209991995 445025017 -1548491841 -1080650855 -1639214278
#>  $ : int [1:7] 10407 -2077260438 -790174784 -16375120 -158426304 1827932304 -891235398
#>  $ : int [1:7] 10407 -550454022 -670204758 -2049815807 -1854005312 580361440 1741761366
#>  $ : int [1:7] 10407 -36707751 -19414859 1887360536 -404830828 474799724 -172578382
#>  $ : int [1:7] 10407 1396806714 1219457057 1178804988 1600756273 290055107 -1994059876
#>  $ : int [1:7] 10407 -1657323527 -1057225955 -1641187400 -102901110 -2076892759 513881516
#>  $ : int [1:7] 10407 284653675 1985606428 177209968 -1907790467 802945999 -581876778
#>  $ : int [1:7] 10407 -1758325729 245858039 466023275 1367257284 -762169796 -847491482
#>  $ : int [1:7] 10407 -821468962 -541193950 1463832331 -1320700965 -1898097791 2025503621
#>  $ : int [1:7] 10407 1845228853 1360181235 294152743 860995603 97516190 602958967
#>  $ : int [1:7] 10407 1228598908 -1567796336 1745638606 -698083358 169397351 -2092615944
#>  $ : int [1:7] 10407 -514971810 -578309597 -545318930 219221692 -1904048266 826868548
#>  $ : int [1:7] 10407 -1796985228 756398136 526841579 364435372 -1625834759 131481095
#>  $ : int [1:7] 10407 1684171695 309362212 35545170 -1926478675 -2093197897 1221798445
#>  $ : int [1:7] 10407 -1131018599 1858804989 -1776889500 -2126576906 562816677 1415546682
#>  $ : int [1:7] 10407 -380475596 1788269262 -882113217 -2029157751 -225534251 560604887
#>  $ : int [1:7] 10407 -708556770 -1231185805 2088200964 1039752396 -639022627 479386290
#>  $ : int [1:7] 10407 629439513 777168086 -968933140 -1215718167 270458685 1418501765
#>  $ : int [1:7] 10407 962125138 969755125 2128382909 -1352644577 78854584 706785326
#>  $ : int [1:7] 10407 -2094839337 -1497291532 753942242 1559205258 1855569001 1515921140
#>  $ : int [1:7] 10407 415854170 -1249195064 303556191 1520352183 406948941 726116134
#>  $ : int [1:7] 10407 -1811231968 1918826163 1170412962 724361528 1461633417 1711491700
#>  $ : int [1:7] 10407 502508376 -1202502881 -1092464878 -75019411 -70658476 -795670526
#>  $ : int [1:7] 10407 -526757661 1308973835 1733875817 -889113106 975059532 -505015098
#>  $ : int [1:7] 10407 -923339428 -1540022806 71934607 691586616 1630173984 -1429921486
#>  $ : int [1:7] 10407 -55239519 698788044 -776238581 496017242 1803552086 -1511856991
#>  $ : int [1:7] 10407 -1562678470 260419472 -520324763 138787047 -110334492 1575204170
#>  $ : int [1:7] 10407 -1024550609 -1759736911 -1736249760 -1841760402 1428288104 1308091309
#>  $ : int [1:7] 10407 485153507 -443013222 -1169355615 168729581 842821683 -113250613
#>  $ : int [1:7] 10407 -205651952 1805613065 -1997872580 -24820130 -1317800898 -1678896428
#>  $ : int [1:7] 10407 -78486210 649758801 255050904 555954344 216793249 -40247846
#>  $ : int [1:7] 10407 674395293 178331363 1390417757 1182623622 -711138290 -1305681389
#>  $ : int [1:7] 10407 1726079310 432028186 1782359479 -1427446075 -1674903020 -1594843267
#>  $ : int [1:7] 10407 872865955 85769091 -1486006396 -447892976 -1272423021 -1696763046
#>  $ : int [1:7] 10407 -751715576 875615152 -949156827 1969921050 -481706255 -1443969781
#>  $ : int [1:7] 10407 845712425 951191754 1418537786 -1824788230 -1646835323 -1481668104
#>  $ : int [1:7] 10407 -784747386 -2080640794 -1331157286 1844285515 1692603415 -303391558
#>  $ : int [1:7] 10407 91462438 1314654688 -1373862637 202703149 -257210769 -331045217
#>  $ : int [1:7] 10407 851041873 1498614739 1571087206 -741324463 -574829083 399271004
#>  $ : int [1:7] 10407 -243558974 164235456 1813247106 -1784978684 -733521041 550308465
#>  $ : int [1:7] 10407 -917611031 1894439019 -1120403642 1147950349 -71658854 460716714
#>  $ : int [1:7] 10407 -1332576236 -60627766 1223671252 1407774992 -1099956096 -1287095972
#>  $ : int [1:7] 10407 1329336202 -77232556 -804338111 523374403 1827366122 401395192
#>  $ : int [1:7] 10407 1730397268 -1503524006 -1970665976 -491998989 -279609277 -1629452228
#>  $ : int [1:7] 10407 -952434051 627163307 -1608595241 958766753 635826770 -1450606507
#>  $ : int [1:7] 10407 396939452 -34484468 -1428702067 278216423 226868897 30922937
#>  $ : int [1:7] 10407 714694092 284581817 201262438 -135595537 230536544 -1782205029
#>  $ : int [1:7] 10407 534502705 -1448331825 -696574553 2121509815 2079726853 1863399970
#>  $ : int [1:7] 10407 -743308048 1058718798 1880753216 -1771145887 2079834728 -1404101868
#>  $ : int [1:7] 10407 -700711710 -2145596900 1780263398 421203698 1012838772 -690830150
#>  $ : int [1:7] 10407 -383919774 1308908104 -329804328 212323831 741504112 -944260307
#>  $ : int [1:7] 10407 1017839099 -1310462786 1944384509 -2144777723 1513733629 59649104
#>  $ : int [1:7] 10407 1732736952 -301617091 -139316597 1334499461 -921228727 -1824969055
#>  $ : int [1:7] 10407 1922764729 -719368083 -874162537 114889506 -1289528325 -925102380
#>  $ : int [1:7] 10407 -1548982484 -1159535824 -949057638 1619002183 -284724109 1635989465
#>  $ : int [1:7] 10407 1711181943 1217720032 1867171318 -510560245 1696638091 -1791710973
#>  $ : int [1:7] 10407 -82334991 -264432056 76737058 137400293 263484684 -1654038047
#>  $ : int [1:7] 10407 188318347 1639427015 -1049603742 -812349004 87907961 -639408741
#>  $ : int [1:7] 10407 1798709238 442346992 -1794747183 -1676648623 -291575864 352288807
#>  $ : int [1:7] 10407 1075693178 812584188 675184382 116958337 363230324 -1251058364
#>  $ : int [1:7] 10407 -1533489853 823042408 1773389861 1886589705 652407422 -2081050220
#>   [list output truncated]