Produce Reproducible Seeds for Parallel Random Number Generation

Produce Reproducible Seeds for Parallel Random Number Generation

make_rng_seeds(count, seed = FALSE)

Arguments

count

The number of RNG seeds to produce.

seed

A logical specifying whether RNG seeds should be generated or not. (seed = NULL corresponds to seed = FALSE). If a list, then it should be of length count and each element should consist of a valid RNG seed.

Value

Returns a non-named list of count independent L'Ecuyer-CMRG random seeds. If seed is NULL or FALSE, then NULL is returned.

Details

This function generates count independent L'Ecuyer-CMRG random seeds that can be used as .Random.seed for parallel processing. These seeds are produced with help of parallel::nextRNGSubStream() and parallel::nextRNGStream() using a strategy that

seed <- <initial RNG seed>
for (ii in seq_len(count)) {
  seeds[[ii]] <- parallel::nextRNGSubStream(seed)
  seed <- parallel::nextRNGStream(seed)
}

This function forwards the RNG state 1 + count times if seed = TRUE.

Examples

## Set up L'Ecuyer-CMRG random seeds for 100 parallel tasks
## based on .Random.seed in the current R process
seeds <- make_rng_seeds(100L, seed = TRUE)
str(seeds)
#> List of 100
#>  $ : int [1:7] 10407 -1067470420 1963645031 1618503100 1908059082 580618691 -1508150195
#>  $ : int [1:7] 10407 -972207856 -1425123879 -1103990543 -1739902797 846667246 -167121226
#>  $ : int [1:7] 10407 1280586354 -1666735509 -29103288 -289869515 1687905019 -257732204
#>  $ : int [1:7] 10407 447677239 -1845594511 97045417 -1656335660 -547565218 -1447852985
#>  $ : int [1:7] 10407 1765894815 386877241 -950962012 -1201207217 -1308625776 437046700
#>  $ : int [1:7] 10407 1269637920 291285316 803607156 -827913989 -1033175160 638838479
#>  $ : int [1:7] 10407 -2134862283 -119571944 -1861235239 -870164988 1959045963 -107906391
#>  $ : int [1:7] 10407 -543895545 -1492449649 383126508 -985674717 -2015850635 -1206088967
#>  $ : int [1:7] 10407 -1915510886 785368375 -935355746 618782931 -1554077343 -1597807761
#>  $ : int [1:7] 10407 289589658 471711571 -1745970745 -1525091834 -1951981553 -704210443
#>  $ : int [1:7] 10407 -799220087 -1166640231 -902906030 414949124 -1975225242 -736026967
#>  $ : int [1:7] 10407 -399959181 2041526153 1883682502 -926401736 -1853627269 -2032112121
#>  $ : int [1:7] 10407 -1660096958 -176808011 -1745846533 900316312 -1179572587 -7677876
#>  $ : int [1:7] 10407 -1515338092 1487897391 -211244973 -440511421 -1627030815 -1360044048
#>  $ : int [1:7] 10407 -408953878 1078544429 -138114536 1845193120 -1322748808 2041289291
#>  $ : int [1:7] 10407 -1927081836 1601856986 1777323516 1254507089 373756424 -924134823
#>  $ : int [1:7] 10407 -1003917076 -583744308 1240244267 -938749681 1194857741 1505363148
#>  $ : int [1:7] 10407 1572514670 -1964740061 -691680757 770043430 -406300625 964513322
#>  $ : int [1:7] 10407 -326483514 -1795002730 402657949 893330745 -1229605537 1255497914
#>  $ : int [1:7] 10407 -836943778 -887558860 -49809074 -1934547055 -661385450 1323558480
#>  $ : int [1:7] 10407 -2013629849 -543054671 -1904203490 1266541483 1413101257 -1341558590
#>  $ : int [1:7] 10407 -1189026115 -1030985058 964845621 1575966434 -700558219 -2029869871
#>  $ : int [1:7] 10407 1009144587 -508301588 190713113 1306261370 1786162093 553379700
#>  $ : int [1:7] 10407 104457173 909603404 1582549161 -1970773561 -1046083098 1176493328
#>  $ : int [1:7] 10407 44698825 -1492800884 -303458603 -1449420029 -2097053672 2022788056
#>  $ : int [1:7] 10407 963512385 617305721 340737728 -1810560430 -697977706 1167748455
#>  $ : int [1:7] 10407 1859033010 1730876389 392290592 969021261 -2034612005 -880670653
#>  $ : int [1:7] 10407 -1694040867 708359964 1380070189 -208839767 1953219178 1232709887
#>  $ : int [1:7] 10407 -1947583395 1304762020 -870165182 -41895749 -839206927 715969566
#>  $ : int [1:7] 10407 -1222682185 -1900420331 2063857017 2065096966 -292021740 151828525
#>  $ : int [1:7] 10407 1329260265 -65528248 70346527 -477402924 -1664252484 -1921707731
#>  $ : int [1:7] 10407 1582818892 1649031255 -1999119222 -1231417813 2120182467 1665349244
#>  $ : int [1:7] 10407 -83600172 134091548 679058307 -924759015 -2091913718 -1279846735
#>  $ : int [1:7] 10407 1661069394 -353493020 -1038213532 -1240198003 278784277 1174266759
#>  $ : int [1:7] 10407 -1297259969 -1877074478 898495431 1769597312 424703457 1750026600
#>  $ : int [1:7] 10407 989630534 1456942875 -864396697 1949541154 -1264127437 121413104
#>  $ : int [1:7] 10407 -867937189 -1617633076 -172897346 385904420 1893036336 -563504451
#>  $ : int [1:7] 10407 1993813389 -1425382262 708500522 -953474431 -1897008186 -2116502878
#>  $ : int [1:7] 10407 2047848012 -1010181887 -180404364 588396925 -818906001 1645053037
#>  $ : int [1:7] 10407 -1764018343 -1358668101 162332680 3060541 -2116254099 755729675
#>  $ : int [1:7] 10407 -2133161962 -916688760 509716609 1898645206 -460464563 -1381913035
#>  $ : int [1:7] 10407 1776257197 -1565265144 -190521891 -105471856 -1886416711 -1679938923
#>  $ : int [1:7] 10407 -134304077 -2025248497 36691722 2132607375 -1431725514 -555514491
#>  $ : int [1:7] 10407 1297595667 1598801228 1121222009 -2033155244 -1019467090 176391283
#>  $ : int [1:7] 10407 1196288853 869035816 -266183538 2061135851 -935159468 -1265340325
#>  $ : int [1:7] 10407 -1121313666 150588475 1328079222 -358098436 -487473580 893633154
#>  $ : int [1:7] 10407 629929387 -587794637 -1704331247 346516283 -410080580 1565489442
#>  $ : int [1:7] 10407 -2099476039 157270330 -2018315872 -231356393 -1753940435 338429963
#>  $ : int [1:7] 10407 -511369565 -1258420294 1442950506 -1452925329 -847525609 1336016551
#>  $ : int [1:7] 10407 1388331225 -1336961030 550103784 556800314 -560660191 -361213439
#>  $ : int [1:7] 10407 1205787491 -1521892810 -981577058 1127903981 -1938000861 1214932751
#>  $ : int [1:7] 10407 1401087241 1444540423 908722503 -595984469 -1166989438 707513803
#>  $ : int [1:7] 10407 -478786627 261680950 -1857370621 1742375814 -1805490472 1547948437
#>  $ : int [1:7] 10407 -1353165940 -1259367274 -1205562854 1453163469 1121505509 -112222453
#>  $ : int [1:7] 10407 -1015122202 -1466196695 2141677199 -378190119 -1617962760 666509008
#>  $ : int [1:7] 10407 1987961262 -1294662613 -947214070 -1833561953 1799541587 1732461437
#>  $ : int [1:7] 10407 1257042805 1978027038 -840928888 313867782 -1752500597 1016818575
#>  $ : int [1:7] 10407 574523812 1685358971 1706392977 1479008721 1647106292 301593903
#>  $ : int [1:7] 10407 1472020814 -2015099324 -705379597 -1442505979 -1325257624 851187058
#>  $ : int [1:7] 10407 -125335179 590643459 -1298484335 1908784255 381553585 1136376574
#>  $ : int [1:7] 10407 67788228 1999360426 -918480062 32385318 -2109704081 82449866
#>  $ : int [1:7] 10407 -151791308 401754748 -622533030 280704970 767503973 798589925
#>  $ : int [1:7] 10407 -1156795003 1657983008 245006089 -1514837397 1943716375 -761783126
#>  $ : int [1:7] 10407 -1323897256 -457485426 2012236616 -1774286352 -891940384 129360754
#>  $ : int [1:7] 10407 -258304940 1664700508 -757639639 -496177248 1991362102 142289026
#>  $ : int [1:7] 10407 -384948767 -511562313 1185986865 80421422 -1849337026 -480708482
#>  $ : int [1:7] 10407 -1027956125 341003835 1564693932 241261462 1530554807 -2006812407
#>  $ : int [1:7] 10407 -361569604 -1748215112 525530982 -1622194242 -1940847438 1103157977
#>  $ : int [1:7] 10407 -900793529 -784900197 1546291362 -835565005 245439937 1720772324
#>  $ : int [1:7] 10407 130281649 755138150 -1206525768 -1370185371 -1797782553 -1938681359
#>  $ : int [1:7] 10407 -1229976326 -347290175 1259892677 1521330376 -733455947 -1067083793
#>  $ : int [1:7] 10407 538479295 62135368 778966519 -171800057 -1490469584 -212126547
#>  $ : int [1:7] 10407 1213833912 -1841914148 -918333985 -606289197 -571986818 2130551998
#>  $ : int [1:7] 10407 1208536603 -755375037 494744860 327071562 1180705561 -740306058
#>  $ : int [1:7] 10407 -1958045931 -1248767057 -1618740676 -246649502 1349151380 1160974644
#>  $ : int [1:7] 10407 910824780 1403518383 -1792285362 41990517 -1132905983 1387529221
#>  $ : int [1:7] 10407 2036522932 -1542639827 1255485229 -1639829812 -718737162 1027014096
#>  $ : int [1:7] 10407 1844018985 -761413692 47167454 553581226 1577279837 -1271924826
#>  $ : int [1:7] 10407 -95625770 -186106027 -73707105 599495680 515284987 -294443627
#>  $ : int [1:7] 10407 515511794 110846461 587717737 -1166707996 777360804 1781809891
#>  $ : int [1:7] 10407 508695951 -1903072693 -2048359577 788792196 -1540903775 603349268
#>  $ : int [1:7] 10407 269802351 -1626665586 29265264 -423679944 -237414541 -813486705
#>  $ : int [1:7] 10407 -454743336 -1117778463 -1833894019 1664866599 2123223565 963493370
#>  $ : int [1:7] 10407 828036977 -1097882151 1817489393 -1069738669 -745669439 -1135978132
#>  $ : int [1:7] 10407 186906664 369007798 -163367493 752059212 155982821 -2065534398
#>  $ : int [1:7] 10407 417141670 1229327634 -950694910 -1777774809 -947225834 1638554956
#>  $ : int [1:7] 10407 1446734209 -1412078077 -429777615 -1348343648 -1798487462 -1790645781
#>  $ : int [1:7] 10407 2067921569 -658778355 -942398452 -555038488 -1264612121 -1721395043
#>  $ : int [1:7] 10407 -1193951996 1268560897 -1436449898 -125562877 1769118637 799922390
#>  $ : int [1:7] 10407 -1807574102 1663365003 558169707 -2039677558 360844735 -486739213
#>  $ : int [1:7] 10407 -1575992294 -54661159 1854311244 -1888885729 696176698 1199869859
#>  $ : int [1:7] 10407 2136836337 -1696371051 99302540 -1951163904 966513194 1107386826
#>  $ : int [1:7] 10407 -1559177232 -192264904 -846757353 592331616 -1227231593 -1053912563
#>  $ : int [1:7] 10407 1094127429 -1537398233 -1709863534 -1984491528 -1406358207 2138461888
#>  $ : int [1:7] 10407 -1670950960 1765206992 1102424541 -1870493735 1082972517 -1954889676
#>  $ : int [1:7] 10407 -1333841926 579479208 1997982003 750140950 -263431736 178245172
#>  $ : int [1:7] 10407 880447765 421234972 1933524048 2110412602 -1516881135 -1261928083
#>  $ : int [1:7] 10407 676536511 -963249057 1650896207 1609542837 371369371 921134290
#>  $ : int [1:7] 10407 1582992218 708582739 -938974908 -754763659 -231609964 -1013643152
#>   [list output truncated]