Produce Reproducible Seeds for Parallel Random Number Generation

Produce Reproducible Seeds for Parallel Random Number Generation

make_rng_seeds(count, seed = FALSE)

Arguments

count

The number of RNG seeds to produce.

seed

A logical specifying whether RNG seeds should be generated or not. (seed = NULL corresponds to seed = FALSE). If a list, then it should be of length count and each element should consist of a valid RNG seed.

Value

Returns a non-named list of count independent L'Ecuyer-CMRG random seeds. If seed is NULL or FALSE, then NULL is returned.

Details

This function generates count independent L'Ecuyer-CMRG random seeds that can be used as .Random.seed for parallel processing. These seeds are produced with help of parallel::nextRNGSubStream() and parallel::nextRNGStream() using a strategy that

seed <- <initial RNG seed>
for (ii in seq_len(count)) {
  seeds[[ii]] <- parallel::nextRNGSubStream(seed)
  seed <- parallel::nextRNGStream(seed)
}

This function forwards the RNG state 1 + count times if seed = TRUE.

Examples

## Set up L'Ecuyer-CMRG random seeds for 100 parallel tasks
## based on .Random.seed in the current R process
seeds <- make_rng_seeds(100L, seed = TRUE)
str(seeds)
#> List of 100
#>  $ : int [1:7] 10407 -821758294 -895574068 -1243721746 871907791 -1634702009 -271926198
#>  $ : int [1:7] 10407 1006502745 163165003 1273367791 1141327001 1018288981 -1802319490
#>  $ : int [1:7] 10407 716835104 -2088656107 188740665 -751613903 628090175 -227712481
#>  $ : int [1:7] 10407 101616600 805356852 -824246180 1325660668 -1041746086 1714541838
#>  $ : int [1:7] 10407 157248083 803739356 -1437973422 1911966029 538284739 1495464982
#>  $ : int [1:7] 10407 -2066883092 1771911578 -1597911031 -601287241 1814880122 544300678
#>  $ : int [1:7] 10407 2012098079 -1810975464 -769105101 699765705 -1734223527 1355590513
#>  $ : int [1:7] 10407 -1934735981 532928370 -1504065640 -187986305 -571270192 -491183399
#>  $ : int [1:7] 10407 -998300691 847581690 -1990661090 -153858005 495099913 897384570
#>  $ : int [1:7] 10407 1867934777 734432405 -313622210 -1312173854 380738959 -124977640
#>  $ : int [1:7] 10407 -1402681631 -1117272950 -1233874211 1315617164 1541741007 364539028
#>  $ : int [1:7] 10407 110333211 -781569893 701986580 -1829101581 -1410017622 1169891010
#>  $ : int [1:7] 10407 -2033390061 1135732680 -232911964 -427595312 -1777857132 -346413001
#>  $ : int [1:7] 10407 293792465 552005252 -565173385 775700400 -1573444802 776923132
#>  $ : int [1:7] 10407 1693839270 -1355549387 2039362270 -294229391 2143430955 -1877222226
#>  $ : int [1:7] 10407 1111803782 -959080625 -1219055835 -1640187528 1476635330 -1943444221
#>  $ : int [1:7] 10407 563201191 1906124946 -922723021 -1999737864 761142941 -1206953107
#>  $ : int [1:7] 10407 1552791108 -1763385349 -1404205194 -1870850630 1947176396 18953675
#>  $ : int [1:7] 10407 2051928557 521343838 923709431 479263896 -1572938637 1481071753
#>  $ : int [1:7] 10407 2076020329 426706721 1686048682 -1699878606 -217412012 -1963051078
#>  $ : int [1:7] 10407 -2127858990 -318021149 -776504248 -1044765396 2140897398 1931259836
#>  $ : int [1:7] 10407 1034795642 27565438 1620215653 422792313 -684098988 838939393
#>  $ : int [1:7] 10407 562796320 -579058275 2071222658 -1258128097 -1596429253 -1517970999
#>  $ : int [1:7] 10407 1864138775 332278363 -2036149746 -541352130 1849847411 -417160925
#>  $ : int [1:7] 10407 824177335 1594908550 660110951 -423685215 1550881831 -1051962433
#>  $ : int [1:7] 10407 -860348624 -2138438388 -253659673 376898651 -1463364036 1156407352
#>  $ : int [1:7] 10407 748076394 232098338 -2024081245 -1081128198 -609365778 -1687701590
#>  $ : int [1:7] 10407 -1500556365 1286798496 1842021407 1452342141 1675713839 -1474381213
#>  $ : int [1:7] 10407 -665079946 -1096370247 -1639649986 164492096 -2025102957 875793677
#>  $ : int [1:7] 10407 -831869622 2119837455 320325014 -1618055555 1119423822 -1736713266
#>  $ : int [1:7] 10407 290671437 1080905898 284586076 -2047012106 1308543249 742972451
#>  $ : int [1:7] 10407 1037297700 468189738 1661318242 -2091399231 -862167865 1884427080
#>  $ : int [1:7] 10407 -1654753669 17213349 -925511895 1847311028 2045171576 -1809668753
#>  $ : int [1:7] 10407 -1772676808 -362482160 2111710758 -299865310 139918467 -238882129
#>  $ : int [1:7] 10407 920240643 -609598338 -1097385156 102268031 -1983078801 1385332442
#>  $ : int [1:7] 10407 -520196453 -827846826 613579099 -230859731 1845457326 1122913123
#>  $ : int [1:7] 10407 -1937006042 -1924568364 -1238658600 1015555767 1537932089 2002912970
#>  $ : int [1:7] 10407 1569121404 1434471650 -443638550 1582203023 -334364018 -532836722
#>  $ : int [1:7] 10407 -482418233 2049630101 327572393 -467788954 -1045052646 -1057696725
#>  $ : int [1:7] 10407 -361878507 133474830 -119858651 -1543054358 1308699764 1752426691
#>  $ : int [1:7] 10407 137667179 776943560 -88979891 589416903 1273987978 -889728137
#>  $ : int [1:7] 10407 -1559426111 905887515 -1900473496 405467637 1794437382 1579819749
#>  $ : int [1:7] 10407 -1468379285 -796567904 -658970248 -107419225 -2061770397 1519065760
#>  $ : int [1:7] 10407 48092017 -673038445 -1284600500 -452495653 1361003074 1590806558
#>  $ : int [1:7] 10407 1739532675 -426176110 709656538 -212704834 1809332282 1097271449
#>  $ : int [1:7] 10407 536933415 -1049830041 -54268684 2026048340 -640383172 1433045978
#>  $ : int [1:7] 10407 -704596870 -760975755 1279075131 133166052 1643714660 -486035098
#>  $ : int [1:7] 10407 812778196 -795031616 -1759119996 -2017020331 429562446 -1766388386
#>  $ : int [1:7] 10407 -1700033317 -1655410236 489968819 -585586299 -122004589 -1303531674
#>  $ : int [1:7] 10407 -1154009132 1706184184 -2107603768 457931549 -2094245701 1236244109
#>  $ : int [1:7] 10407 -487536684 377613035 1223073963 -886548239 1163864932 732872984
#>  $ : int [1:7] 10407 -811289662 1187619194 593724424 -644031205 393993718 -1429147447
#>  $ : int [1:7] 10407 179341714 684500216 -2124974665 -1188394364 398343440 872141699
#>  $ : int [1:7] 10407 -1496303047 190335330 2012665125 -593315813 1903084801 1864537742
#>  $ : int [1:7] 10407 -775448511 1498832924 -1113315015 1448438570 602001770 2040138166
#>  $ : int [1:7] 10407 -13913004 -2012645632 -1580117318 -116451831 531468632 -863242584
#>  $ : int [1:7] 10407 1290410764 -199486678 -1264861889 -1256425761 -212467888 -2130022202
#>  $ : int [1:7] 10407 1126843544 715046615 424061986 587591256 1813056988 1975530871
#>  $ : int [1:7] 10407 -1470988183 -729924016 -78553056 -1867708027 -1066955490 -2094089053
#>  $ : int [1:7] 10407 1159907829 36328342 -15390515 -1163537037 -786950061 1804418577
#>  $ : int [1:7] 10407 -143687666 -1846537163 1514423815 -1283268317 -931301693 1427294613
#>  $ : int [1:7] 10407 1934715597 -374230399 -412696529 111899877 -54936870 -1605601235
#>  $ : int [1:7] 10407 2055124134 -1599388541 1725890762 194500368 1042499030 759629390
#>  $ : int [1:7] 10407 -212660815 -1154214722 -1541077136 1902569585 1921230863 -243418879
#>  $ : int [1:7] 10407 -421096957 -564235065 -1510034518 1529858326 -326553054 1935233756
#>  $ : int [1:7] 10407 -1347770667 -1155210961 -1967470501 -1403397591 1752006287 -2051202731
#>  $ : int [1:7] 10407 -1131258089 -254683777 1012810242 2028529069 -1333003298 -827206272
#>  $ : int [1:7] 10407 328342125 1930670878 1810228941 1223283981 758206155 1917244168
#>  $ : int [1:7] 10407 -1222783951 -641844277 98280441 -68136628 1984947317 1985093798
#>  $ : int [1:7] 10407 2005137938 -1958840581 -1434508941 205640532 2044722374 52259850
#>  $ : int [1:7] 10407 -1350159137 -1582012674 684984828 -1542100644 -330964946 -2048592001
#>  $ : int [1:7] 10407 -2091129594 963803578 497887151 1428309860 -1458691976 -497526528
#>  $ : int [1:7] 10407 1374872731 -183511198 1606896066 1060100235 1357021308 -809147696
#>  $ : int [1:7] 10407 1437424733 -863484168 319891167 141925861 -832047495 -1708529839
#>  $ : int [1:7] 10407 -397547782 -40253127 -1109509235 -417151430 1175100051 -928357064
#>  $ : int [1:7] 10407 544900863 1792112389 -1141878765 -367974499 -1795626198 -707033609
#>  $ : int [1:7] 10407 1812851936 1269344018 -623562734 648658314 -601271497 2036305195
#>  $ : int [1:7] 10407 -1194929702 -2007141291 -988209858 -690368634 -730195003 369621559
#>  $ : int [1:7] 10407 -778310466 1380037804 919799030 -1404931061 -1082817379 -1382733521
#>  $ : int [1:7] 10407 1205259234 496189021 1304928162 1069564904 757274644 -1602267423
#>  $ : int [1:7] 10407 99898586 -1220673567 575835318 -1195939665 -1688072985 -1616834289
#>  $ : int [1:7] 10407 981947855 -1553451526 -59949636 -921733399 636148671 -284502549
#>  $ : int [1:7] 10407 -1460932168 1785463504 -173446066 -288131591 -1646537060 727845724
#>  $ : int [1:7] 10407 727375919 1615471081 -502057370 -661273811 -216542192 -2128973562
#>  $ : int [1:7] 10407 1931467100 -1912864233 -1940449559 138507544 1155956634 2062080045
#>  $ : int [1:7] 10407 1911746367 742577942 -6600917 980090063 -848968530 -1694906304
#>  $ : int [1:7] 10407 -595412310 -1391208337 1350544872 -2008585042 1810151719 1133874494
#>  $ : int [1:7] 10407 -132493534 -1510024524 -510347776 1947233115 1344808920 607185235
#>  $ : int [1:7] 10407 1454857529 -583002692 84085181 -698218065 544108278 -402355730
#>  $ : int [1:7] 10407 620449438 1777677110 369241981 1121564392 471216124 1788810360
#>  $ : int [1:7] 10407 1225384274 -1859725907 -438879463 -97252147 -1636984589 -2127143723
#>  $ : int [1:7] 10407 -24166193 1386047796 -1322959869 -322656518 522781399 -1372544754
#>  $ : int [1:7] 10407 1636042211 1459540407 1568419024 1812302911 -1156369414 -418425865
#>  $ : int [1:7] 10407 2111375792 -757158837 1885791284 -396173224 -1295279520 -1297609202
#>  $ : int [1:7] 10407 -589931857 -1572853187 525535644 802182618 -632683239 -1255385645
#>  $ : int [1:7] 10407 101011665 -388756256 597115413 -1102032578 906644310 1608036897
#>  $ : int [1:7] 10407 1960469023 -1866343230 -1809905988 -623083873 385170271 -170238815
#>  $ : int [1:7] 10407 1741187094 287226548 -1349881285 -420536089 -2095531905 733513964
#>  $ : int [1:7] 10407 1344205556 930819322 -1561061394 -897582269 1904629954 -29708653
#>   [list output truncated]