Produce Reproducible Seeds for Parallel Random Number Generation

Produce Reproducible Seeds for Parallel Random Number Generation

make_rng_seeds(count, seed = FALSE)

Arguments

count

The number of RNG seeds to produce.

seed

A logical specifying whether RNG seeds should be generated or not. (seed = NULL corresponds to seed = FALSE). If a list, then it should be of length count and each element should consist of a valid RNG seed.

Value

Returns a non-named list of count independent L'Ecuyer-CMRG random seeds. If seed is NULL or FALSE, then NULL is returned.

Details

This function generates count independent L'Ecuyer-CMRG random seeds that can be used as .Random.seed for parallel processing. These seeds are produced with help of parallel::nextRNGSubStream() and parallel::nextRNGStream() using a strategy that

seed <- <initial RNG seed>
for (ii in seq_len(count)) {
  seeds[[ii]] <- parallel::nextRNGSubStream(seed)
  seed <- parallel::nextRNGStream(seed)
}

This function forwards the RNG state 1 + count times if seed = TRUE.

Examples

## Set up L'Ecuyer-CMRG random seeds for 100 parallel tasks
## based on .Random.seed in the current R process
seeds <- make_rng_seeds(100L, seed = TRUE)
str(seeds)
#> List of 100
#>  $ : int [1:7] 10407 -1310251811 -1874489875 1169293762 454515851 1704929781 1162643240
#>  $ : int [1:7] 10407 -1676804425 -1459245750 -1922302586 -752269354 -634909601 1840441902
#>  $ : int [1:7] 10407 -88031724 -655332874 -1378618449 -585541381 -1933297133 -419212423
#>  $ : int [1:7] 10407 -1771615988 1643779560 470015228 1979831003 295532058 -331725021
#>  $ : int [1:7] 10407 -2111869621 158862512 -745899488 -644150535 246852697 1199384419
#>  $ : int [1:7] 10407 -139168454 1226192231 1990448148 -1722163681 -53797775 629784691
#>  $ : int [1:7] 10407 11641622 -1931830783 2113337124 374423556 977743063 -273728792
#>  $ : int [1:7] 10407 44428171 -858221913 298577497 -689718142 -1073412989 178423791
#>  $ : int [1:7] 10407 -1045482006 68425425 -1185808313 -1747534414 -424278240 -1474099605
#>  $ : int [1:7] 10407 -91250210 -1453409469 284787535 -360668683 -1964665940 -1272542923
#>  $ : int [1:7] 10407 -161470837 -1762136913 1027646865 -1260438946 -684508625 -1060610658
#>  $ : int [1:7] 10407 -1561803331 -984644730 51538731 931004823 -1974176025 -2079376759
#>  $ : int [1:7] 10407 -153255418 1055981577 1980817243 -1804290043 416052720 1382559778
#>  $ : int [1:7] 10407 1596115636 2027052698 -456458673 -2013738899 -1978858809 -1199158537
#>  $ : int [1:7] 10407 186193615 -1470535216 -186067049 695284454 715348803 255813642
#>  $ : int [1:7] 10407 529837909 -2096984814 -1434299917 -1771875782 1917959178 -773451568
#>  $ : int [1:7] 10407 -574746364 54344521 353291732 -1955680836 -1192751739 893219968
#>  $ : int [1:7] 10407 895573057 -375956994 1952534999 1144239636 -910596154 1794206968
#>  $ : int [1:7] 10407 1995128836 1429808968 1796794426 -2066000530 -1789659602 -2101114927
#>  $ : int [1:7] 10407 -875189323 -1343414718 1747763157 -1729629117 548778151 -1870138067
#>  $ : int [1:7] 10407 1036395661 -1279399616 -1617614325 1580726700 -570450300 1645883092
#>  $ : int [1:7] 10407 2001695497 1661639737 1260722211 826956857 1500143532 1904582928
#>  $ : int [1:7] 10407 1239331642 -562920414 1999020478 -1892399681 2027251927 -767345142
#>  $ : int [1:7] 10407 -1280097045 777859734 180666839 1754611433 1876058914 -1462486810
#>  $ : int [1:7] 10407 -89237689 387767382 569312414 2057840545 1913750749 -249411729
#>  $ : int [1:7] 10407 1689308610 -1038416746 -651098815 472854540 -639182969 2091222898
#>  $ : int [1:7] 10407 396374042 836965020 930847204 863053252 1262968130 1617188878
#>  $ : int [1:7] 10407 -1781364204 44974280 339974370 1539419958 2131784529 -1849419231
#>  $ : int [1:7] 10407 2001388060 -1676094291 626878852 -1116342166 26731225 -1374632775
#>  $ : int [1:7] 10407 -822257168 522451860 -1476304003 -1755884379 1728893908 -1101533625
#>  $ : int [1:7] 10407 -1694568106 -983058540 847719498 -1684723347 -1787043976 357973312
#>  $ : int [1:7] 10407 -652653128 518893817 -1560240991 591206526 1099643988 1184817662
#>  $ : int [1:7] 10407 -1940491906 21240676 -147109276 775158950 -1877383073 -1899125461
#>  $ : int [1:7] 10407 -1022359423 -145508415 622438325 296128889 1093331522 -2088820852
#>  $ : int [1:7] 10407 1927967569 -333930345 -630770263 -1810410319 -1971064798 -544741366
#>  $ : int [1:7] 10407 1958754431 -1857139359 -1849886820 -1563133292 1053926748 -573622248
#>  $ : int [1:7] 10407 1337581282 -1592552912 295592496 934568176 -1640343600 -1652928951
#>  $ : int [1:7] 10407 153086602 -1810830127 -727085332 -1243256042 1311103592 -1395465602
#>  $ : int [1:7] 10407 899164649 1291622614 -589285661 -1151234772 715026555 -100784929
#>  $ : int [1:7] 10407 1079424521 -1835596208 2017901576 375786961 1448294751 901560513
#>  $ : int [1:7] 10407 274262531 1814670595 -1904723496 -1620599627 -1359205779 -1060927437
#>  $ : int [1:7] 10407 529844230 2102329443 -2027973905 1503575380 -1358572570 729964335
#>  $ : int [1:7] 10407 -912335949 1553812527 1520131254 -1752271834 1598159060 -1591700687
#>  $ : int [1:7] 10407 635292089 1721020299 993376659 1577591511 1526106893 -1639863754
#>  $ : int [1:7] 10407 -1119824253 1851230111 1479695722 -1575378185 1937508101 -552176977
#>  $ : int [1:7] 10407 744770270 -1405616047 -1562368560 341561299 -1441983574 -717725586
#>  $ : int [1:7] 10407 -1921117034 1710405529 1013521867 316734102 1092532025 -1731887337
#>  $ : int [1:7] 10407 -260641812 1743368717 1961890131 -250672122 -1716472949 -164612178
#>  $ : int [1:7] 10407 1109310331 142721629 -192146907 -912851 -124556880 1801290193
#>  $ : int [1:7] 10407 1971429411 -207494447 195002096 2096414030 -369296276 -458662279
#>  $ : int [1:7] 10407 -286107132 2083902238 -505317293 2092702990 -1949030693 -358329252
#>  $ : int [1:7] 10407 -1779284722 1405273915 -1980359075 351284753 -630332028 -366365037
#>  $ : int [1:7] 10407 937690053 -2082212004 -196881446 1090613021 1186056622 -2022455320
#>  $ : int [1:7] 10407 -845656914 -750005228 -1706525433 1373128372 948506564 -1802536784
#>  $ : int [1:7] 10407 960940150 -1060084910 -1390010888 -1911353640 1149486585 1079355401
#>  $ : int [1:7] 10407 -843308477 795019011 -1448228718 -1396434717 1072723338 1141326298
#>  $ : int [1:7] 10407 344331274 902213852 -1548080802 532326217 -794293123 938394977
#>  $ : int [1:7] 10407 -521909706 -2007817839 -620092208 1111501258 1088142848 -646181524
#>  $ : int [1:7] 10407 -1694972377 -773989727 1705576032 -1771613218 1435765935 -441242736
#>  $ : int [1:7] 10407 1114402081 2116310598 195721881 1666911400 470475517 1384004866
#>  $ : int [1:7] 10407 1570445002 200362068 -2105913936 -981942622 -444694701 1720604214
#>  $ : int [1:7] 10407 1447161156 -1389496613 1000719553 23877829 2032199951 1465470208
#>  $ : int [1:7] 10407 -475530110 -572818717 947973550 920230338 -220974581 -473307790
#>  $ : int [1:7] 10407 -1050890878 -1880631104 -1131833592 -824840700 -628750372 -1697731970
#>  $ : int [1:7] 10407 871661233 587800551 165012916 -847290561 253311045 -870840144
#>  $ : int [1:7] 10407 -207532226 1759262729 -151168742 -263864510 1377598246 1458645298
#>  $ : int [1:7] 10407 -1311523712 -929368488 -342254928 -402227010 -963546786 1970116799
#>  $ : int [1:7] 10407 39765775 -476508145 -508324389 -2081661081 1155641186 -388016336
#>  $ : int [1:7] 10407 1899041718 -1345282840 659357874 1284073656 428632892 1030051121
#>  $ : int [1:7] 10407 -1478975175 -575445022 772843185 820566508 1365807685 -1618954481
#>  $ : int [1:7] 10407 -1775998100 92099788 740835274 1935057801 -818982186 1891201124
#>  $ : int [1:7] 10407 1784775614 -1388744836 257710076 -946650169 -1072828347 -954185681
#>  $ : int [1:7] 10407 -1467564942 -531923167 -296012472 -1171594686 -1641966401 2044109002
#>  $ : int [1:7] 10407 -451333871 -112820987 1109348875 -540013386 -711475188 -393781347
#>  $ : int [1:7] 10407 1392183981 534028171 -429530536 -2047614004 1714333472 1804726128
#>  $ : int [1:7] 10407 1736171573 747089410 1966857431 1120576770 -1935818561 -2133690157
#>  $ : int [1:7] 10407 -506946017 1730965604 -1472857223 906255364 1066102729 805819614
#>  $ : int [1:7] 10407 -814701304 -491602908 1020269767 720694488 111378924 -330750745
#>  $ : int [1:7] 10407 1433749336 -1445256129 343449690 1108289815 -2095994099 2076282442
#>  $ : int [1:7] 10407 -967046017 -1608272365 2064223577 -1518654084 1941676865 221209752
#>  $ : int [1:7] 10407 -2142062443 -1137900916 1360545046 1512401613 -1455156158 186606286
#>  $ : int [1:7] 10407 498399693 -1929973450 612611860 1797097314 -91623212 -25144423
#>  $ : int [1:7] 10407 1818621867 1381401807 -616531870 -71365263 -1192495721 2019463518
#>  $ : int [1:7] 10407 48238645 706900885 1817562906 -177562826 -1999063260 1814051819
#>  $ : int [1:7] 10407 1216814833 -73858787 -361489043 1383549356 1485038737 -1290431942
#>  $ : int [1:7] 10407 -569323575 -1971375252 809962812 -1345814143 1875173553 -312283533
#>  $ : int [1:7] 10407 -2083430468 313936976 1211220907 -922700934 964145005 872286819
#>  $ : int [1:7] 10407 -746850395 634949421 -980547501 -99722004 78185660 -2118417754
#>  $ : int [1:7] 10407 66351433 1469105030 -2111613583 -1615558651 -294800778 988941396
#>  $ : int [1:7] 10407 527385042 1446251585 978915068 55483317 407838882 -588788076
#>  $ : int [1:7] 10407 -1240942975 -895424646 -1303435554 1719700459 1843132147 1615747440
#>  $ : int [1:7] 10407 2119424605 -2029571594 -925850204 -1376659922 -1014362799 -1457323717
#>  $ : int [1:7] 10407 801237819 -1896070368 368451415 420310824 1059733445 -487404033
#>  $ : int [1:7] 10407 472498506 -1775805167 -549334485 -1183372450 1948135032 -469819446
#>  $ : int [1:7] 10407 -1717488738 -7417295 756808640 1608988687 999291554 -1382920146
#>  $ : int [1:7] 10407 1069169619 1943339465 1402923437 808710138 -317084026 1304018325
#>  $ : int [1:7] 10407 1787143326 699416102 -1671479266 -670872822 -2112569952 1237301368
#>  $ : int [1:7] 10407 -1833099431 -240028174 144866627 872346356 -793358615 1107064884
#>  $ : int [1:7] 10407 -1966229022 -996142297 -538845320 548055345 -954639214 279596005
#>   [list output truncated]