Produce Reproducible Seeds for Parallel Random Number Generation

Produce Reproducible Seeds for Parallel Random Number Generation

make_rng_seeds(count, seed = FALSE)

Arguments

count

The number of RNG seeds to produce.

seed

A logical specifying whether RNG seeds should be generated or not. (seed = NULL corresponds to seed = FALSE). If a list, then it should be of length count and each element should consist of a valid RNG seed.

Value

Returns a non-named list of count independent L'Ecuyer-CMRG

random seeds. If seed is NULL or FALSE, then NULL is returned.

Details

This function generates count independent L'Ecuyer-CMRG random seeds that can be used as .Random.seed for parallel processing. These seeds are produced with help of parallel::nextRNGSubStream() and parallel::nextRNGStream() using a strategy that

seed <- <initial RNG seed>
for (ii in seq_len(count)) {
  seeds[[ii]] <- parallel::nextRNGSubStream(seed)
  seed <- parallel::nextRNGStream(seed)
}

This function forwards the RNG state 1 + count times if seed = TRUE.

Examples

## Set up L'Ecuyer-CMRG random seeds for 100 parallel tasks
## based on .Random.seed in the current R process
seeds <- make_rng_seeds(100L, seed = TRUE)
str(seeds)
#> List of 100
#>  $ : int [1:7] 10407 483017091 285091851 1221257312 1540084244 1787178713 -834287317
#>  $ : int [1:7] 10407 728469611 -1892083221 -1100917497 841570262 1972444494 1956313599
#>  $ : int [1:7] 10407 -1870641104 -694075988 -1304613259 -1732550590 -1475136892 -1221732071
#>  $ : int [1:7] 10407 623081151 -2137204960 1227525037 1011037145 -1858823377 74712463
#>  $ : int [1:7] 10407 1316128831 -598743000 530020694 830611656 460765114 -1654070636
#>  $ : int [1:7] 10407 -37847940 215984944 -746430935 1912049234 441878747 68587007
#>  $ : int [1:7] 10407 -544512764 951091801 -1559019111 -1020839893 62794685 -599417392
#>  $ : int [1:7] 10407 1303712349 1709922587 369175385 -31253189 445954033 1393322694
#>  $ : int [1:7] 10407 1660461476 -1906175890 -1887287806 -1564057272 1596573010 119816213
#>  $ : int [1:7] 10407 -2101562487 1904634938 -1891949539 -1440952346 -1657538637 928966795
#>  $ : int [1:7] 10407 -598135917 701089509 1133758446 -154450462 483676457 -1824769155
#>  $ : int [1:7] 10407 -732233928 1557913758 -1531584004 -1818542799 359947980 -383944614
#>  $ : int [1:7] 10407 -1183944409 -666895830 -71047501 524211091 -322201310 925523231
#>  $ : int [1:7] 10407 -1667291214 1500789473 2042603384 -332160193 856718614 -2145388866
#>  $ : int [1:7] 10407 -1971338318 277218469 1177111777 -1969802889 1778782000 825971987
#>  $ : int [1:7] 10407 -752175999 938287068 -734108281 118593987 -1408902249 396347128
#>  $ : int [1:7] 10407 1700943273 447471708 1875931286 -1815613108 375345380 -1564637481
#>  $ : int [1:7] 10407 -1137938062 1647822560 -596553137 1297020429 1152785822 -585522300
#>  $ : int [1:7] 10407 1354533669 -960747027 1434168226 1576520033 1810386549 -1409231740
#>  $ : int [1:7] 10407 -856566329 -1185017198 1197942572 -1863800538 1286413274 1638808124
#>  $ : int [1:7] 10407 -1071768430 -576754469 -1152510638 -813775124 828699201 1882003877
#>  $ : int [1:7] 10407 2096221774 -222065428 -1523737541 1701837066 -667333972 1716532711
#>  $ : int [1:7] 10407 1531143153 -691057720 808521651 2096910965 406338512 -1421607209
#>  $ : int [1:7] 10407 536197541 5636495 -308950534 918178815 1152198415 1278951507
#>  $ : int [1:7] 10407 -1704597800 -1295619721 1256133911 -1744520201 -1653984034 775725926
#>  $ : int [1:7] 10407 1822488147 -1311503178 4685313 -2065754514 -795920848 208943386
#>  $ : int [1:7] 10407 -1720026422 704082246 1569570072 -547974621 697241915 1133645034
#>  $ : int [1:7] 10407 -1620653299 -832970841 331859004 1045180513 636987307 872757275
#>  $ : int [1:7] 10407 -1388983368 -1628878888 893568753 -35019392 -1069356359 745688992
#>  $ : int [1:7] 10407 1227645906 419485844 1540698878 373345936 492643290 1951757346
#>  $ : int [1:7] 10407 -722608395 -1125356538 27640352 -1897308520 -911404072 -735144369
#>  $ : int [1:7] 10407 10720196 -1906342009 1553930820 -1218112404 999810772 1117405142
#>  $ : int [1:7] 10407 -834653309 115295983 2050109926 1318753406 1710186560 812554459
#>  $ : int [1:7] 10407 1621647042 -963061247 -81741107 -1136314697 651465269 -2104658961
#>  $ : int [1:7] 10407 -1562231988 -1902919452 1097753196 667590962 -431890792 2001953441
#>  $ : int [1:7] 10407 -1835855129 -1312300864 -703718401 1849872218 -337683719 -2040616604
#>  $ : int [1:7] 10407 973506720 590363456 -1389533399 1156878785 -1722416604 1791546051
#>  $ : int [1:7] 10407 -1932983088 -517166589 1120523054 1116949922 1809038396 -1180939478
#>  $ : int [1:7] 10407 361339809 487301047 -1909773258 -1619325093 -1593620244 -1122199691
#>  $ : int [1:7] 10407 -1328830864 -790757965 1398183132 857471007 6405412 2134984500
#>  $ : int [1:7] 10407 1274955311 -231209175 -1745417690 -1344805681 -1499645926 232484807
#>  $ : int [1:7] 10407 -13902918 187049836 -2124033761 -335701508 720880252 -372107880
#>  $ : int [1:7] 10407 -862327300 -842298312 21911587 1683842413 -113464688 980979689
#>  $ : int [1:7] 10407 1950205486 -2058215975 -1132469746 -816802770 2145185451 252620697
#>  $ : int [1:7] 10407 -596199415 598911403 1719643089 1101508532 -281300907 -266157353
#>  $ : int [1:7] 10407 -2083658046 1080934996 1102034606 1419928818 -1667200269 -1707331781
#>  $ : int [1:7] 10407 -1352958048 234966440 528206302 2053017173 -383044591 2127578758
#>  $ : int [1:7] 10407 645452125 1439433944 -1042900393 268055096 2134889450 -772771394
#>  $ : int [1:7] 10407 -845791452 -689123127 -821941799 689615828 -1439231858 965516113
#>  $ : int [1:7] 10407 -2063161946 -1072402121 -1768022050 -1213345736 -2130424465 850537005
#>  $ : int [1:7] 10407 -1764597716 286376668 -1666994297 1650959243 470503560 -1203360421
#>  $ : int [1:7] 10407 1281152254 304194436 1287080465 1911513015 1359902511 324621856
#>  $ : int [1:7] 10407 1872665208 -790140886 996456472 -835122711 225320057 -1645740574
#>  $ : int [1:7] 10407 337377840 77967465 1635844403 1407707627 -1147055521 1309037204
#>  $ : int [1:7] 10407 1508800020 -1214058782 1399685574 933772196 1088399086 -447430984
#>  $ : int [1:7] 10407 807148933 -1048780929 -27481816 -1035111245 -319680003 -838926111
#>  $ : int [1:7] 10407 -515954499 1978288398 -1460305875 -1948814079 -1835122203 1410141604
#>  $ : int [1:7] 10407 -1205830205 -2070088103 1387697661 -1710658655 -1784358312 -1958942611
#>  $ : int [1:7] 10407 -1782011196 1234186869 -1968826252 48376090 -605716958 1557887435
#>  $ : int [1:7] 10407 182831049 -1272240156 -1889427630 617478526 -1053741247 1845682610
#>  $ : int [1:7] 10407 -1706062445 -442950392 -439895818 -1840793717 1141092462 -585289835
#>  $ : int [1:7] 10407 63791975 -1733940939 36568335 903598385 146949671 1442765193
#>  $ : int [1:7] 10407 -1006389314 1279424998 -1157434028 -753234986 -502411748 -613608485
#>  $ : int [1:7] 10407 1362181345 -774129196 759781240 -1921649583 -1178361326 690292188
#>  $ : int [1:7] 10407 -1822300832 1345794807 474663367 -1553736708 1465270944 -1840291291
#>  $ : int [1:7] 10407 -36821947 928777598 -1268991003 32168319 200235371 562415356
#>  $ : int [1:7] 10407 -936698073 1846736625 -348229758 -651544482 -238447235 1016379590
#>  $ : int [1:7] 10407 1289521645 1458625185 -852768712 -2063224264 -1439468264 507418203
#>  $ : int [1:7] 10407 -444893187 490393530 1545626730 1914044030 -1501823800 -1742755079
#>  $ : int [1:7] 10407 1908237631 -500410584 -1613486902 -384898418 -1441629529 57966904
#>  $ : int [1:7] 10407 1234049186 1838679948 122999555 387714732 -1361701679 -2069153319
#>  $ : int [1:7] 10407 -1365192629 -1939952803 1819752577 119444950 462548151 -670250766
#>  $ : int [1:7] 10407 173698862 -458063769 -936113636 1838543381 -1280909713 2050901509
#>  $ : int [1:7] 10407 393285984 -598906659 -1328064783 -128078796 1621115384 839819554
#>  $ : int [1:7] 10407 -235156004 -952790970 -1270244198 -27027207 1977864679 -818145516
#>  $ : int [1:7] 10407 -1419467204 -1395169739 611110088 -445504043 -1299592298 -1061048299
#>  $ : int [1:7] 10407 -1909948531 -912848647 -2134100424 -399311361 1263944005 -979728902
#>  $ : int [1:7] 10407 -1119591024 488858275 1307913801 679682180 -797187031 -942335448
#>  $ : int [1:7] 10407 -381181195 1268320434 161170010 280586916 -2023745577 776653265
#>  $ : int [1:7] 10407 1518368949 1066496109 609737176 296630818 1895459369 -2103659876
#>  $ : int [1:7] 10407 -1342063790 -335496945 1062494510 429336647 -2022688385 2114108635
#>  $ : int [1:7] 10407 1627489536 -962953672 -488880204 -1459419440 -309131472 1469016243
#>  $ : int [1:7] 10407 152323802 1855572098 -1732545899 2101637674 1094211729 -556589199
#>  $ : int [1:7] 10407 1103909540 830907386 -693394477 -82414556 -1983197993 -2065481073
#>  $ : int [1:7] 10407 -1959702949 -167551177 245768245 -1490378925 1276663543 -1356123872
#>  $ : int [1:7] 10407 -1630120512 -711930139 -655343456 1790211515 -1585768268 -1601005248
#>  $ : int [1:7] 10407 1376707940 -494106051 337746014 1454460081 -1466280891 1762471936
#>  $ : int [1:7] 10407 1123935854 2010221285 1760860809 -1682461426 -933909584 1090749456
#>  $ : int [1:7] 10407 1949169028 1382328649 462096115 362276455 690441442 -713673092
#>  $ : int [1:7] 10407 1050392286 1336859158 -969125246 -888357008 -1392107662 -276173947
#>  $ : int [1:7] 10407 759875272 1212805160 175997517 -1451607566 1309086107 -441176227
#>  $ : int [1:7] 10407 1124079211 -1633138426 1976563771 -1018522847 -741661157 -782686204
#>  $ : int [1:7] 10407 -174582770 -2102882661 182469952 1703644890 -579308425 -1573723675
#>  $ : int [1:7] 10407 474756608 -1266676685 -1686730689 1607766079 -2066026039 -1285502328
#>  $ : int [1:7] 10407 -330294005 -746384180 -276643081 342973372 -1783565300 2067608538
#>  $ : int [1:7] 10407 -1529965389 1368363553 1055881036 -1121578369 2077206763 -1856296202
#>  $ : int [1:7] 10407 -1857206249 1915551600 -1193722159 -1086245809 -1287112155 1860439457
#>  $ : int [1:7] 10407 1525638658 -1791225756 155314123 134870851 784961667 -19521069
#>  $ : int [1:7] 10407 -1486879585 334649903 -700877776 505813805 1109538045 424999325
#>   [list output truncated]