Produce Reproducible Seeds for Parallel Random Number Generation

Produce Reproducible Seeds for Parallel Random Number Generation

make_rng_seeds(count, seed = FALSE)

Arguments

count

The number of RNG seeds to produce.

seed

A logical specifying whether RNG seeds should be generated or not. (seed = NULL corresponds to seed = FALSE). If a list, then it should be of length count and each element should consist of a valid RNG seed.

Value

Returns a non-named list of count independent L'Ecuyer-CMRG

random seeds. If seed is NULL or FALSE, then NULL is returned.

Details

This function generates count independent L'Ecuyer-CMRG random seeds that can be used as .Random.seed for parallel processing. These seeds are produced with help of parallel::nextRNGSubStream() and parallel::nextRNGStream() using a strategy that

seed <- <initial RNG seed>
for (ii in seq_len(count)) {
  seeds[[ii]] <- parallel::nextRNGSubStream(seed)
  seed <- parallel::nextRNGStream(seed)
}

This function forwards the RNG state 1 + count times if seed = TRUE.

Examples

## Set up L'Ecuyer-CMRG random seeds for 100 parallel tasks
## based on .Random.seed in the current R process
seeds <- make_rng_seeds(100L, seed = TRUE)
str(seeds)
#> List of 100
#>  $ : int [1:7] 10407 967854838 -1104804427 1151927601 1389239914 342485256 -2047654148
#>  $ : int [1:7] 10407 1675720948 -1444004700 1064545105 -1594470889 -1800930556 1545874407
#>  $ : int [1:7] 10407 -1737184466 -536014582 -816910969 -316793082 -429290755 1180583342
#>  $ : int [1:7] 10407 -514210288 704556051 1236550613 1057486627 -1873174653 94653185
#>  $ : int [1:7] 10407 438013129 892399161 -593850426 460315863 -1157990190 1061140355
#>  $ : int [1:7] 10407 -2147125706 2056443138 1631940997 1446785069 -83306227 1507249307
#>  $ : int [1:7] 10407 1188202329 1224130294 684476339 1211592661 -926372424 153905798
#>  $ : int [1:7] 10407 -108543136 -815676501 -28592196 -727835587 455833609 1991158926
#>  $ : int [1:7] 10407 -2038392620 788966021 210527736 -1805961275 -1963952166 1834090931
#>  $ : int [1:7] 10407 -1909171592 1428575001 321111618 -421145496 1822363784 615724594
#>  $ : int [1:7] 10407 -853584956 -1304516870 1159532832 -1358676209 1220185520 -1579217896
#>  $ : int [1:7] 10407 -274836729 1755859550 -1986484162 480662729 1253443528 -1075044367
#>  $ : int [1:7] 10407 954319679 -934421529 1568728210 -1240406267 -2108443437 -1781587411
#>  $ : int [1:7] 10407 -1278148797 195783891 528693848 -800878605 -1871953691 695353432
#>  $ : int [1:7] 10407 -151433682 -116646870 839695660 871426962 364253321 -1839683352
#>  $ : int [1:7] 10407 -1980951420 1744896052 702844463 624493898 -1765774312 900801534
#>  $ : int [1:7] 10407 745766817 265424102 -1805044324 -57156420 -1042812409 739200228
#>  $ : int [1:7] 10407 965266942 1444275253 -1352280158 -694041398 -1979835362 247159584
#>  $ : int [1:7] 10407 1484724434 -1198229625 1846432297 -1872136950 -1193622844 -886727260
#>  $ : int [1:7] 10407 526386521 -1419272917 226796484 -1346910728 1733465102 1656867723
#>  $ : int [1:7] 10407 251794709 -1173134349 -938240077 1311840390 1697725996 703947688
#>  $ : int [1:7] 10407 -1298342578 -445001476 1928432178 996336491 721410732 -163776067
#>  $ : int [1:7] 10407 -1448831714 -1556044191 1244214430 629827762 -1612041476 -2084412214
#>  $ : int [1:7] 10407 236338366 1317436110 1858177135 -1819416773 1876628518 -602826242
#>  $ : int [1:7] 10407 398775360 1890352276 -1681222087 -163716558 -1975183167 766585864
#>  $ : int [1:7] 10407 585396215 -1189309895 848434981 -1927315143 2019746621 -2001213520
#>  $ : int [1:7] 10407 1701071750 893738460 -862200110 142246433 2049961759 -1334022109
#>  $ : int [1:7] 10407 629580527 609043191 1904131196 -170054024 -1434113105 -734467126
#>  $ : int [1:7] 10407 345120029 -413152287 -246612373 -899802271 1441211055 -1137091107
#>  $ : int [1:7] 10407 -112451461 1775474117 -1890201532 -1945462967 -430550920 -945355520
#>  $ : int [1:7] 10407 -224079763 580939235 -1096921945 -1111711211 -483462273 553047173
#>  $ : int [1:7] 10407 -614332655 1468763951 1037955138 -309336651 1606886839 -1273445939
#>  $ : int [1:7] 10407 -769810508 2144222266 1889860088 1250782928 804815497 -1134537295
#>  $ : int [1:7] 10407 938550027 -360935759 1435493025 972547323 37209792 -31688638
#>  $ : int [1:7] 10407 1679813134 601678466 -1168456221 -863386676 752502607 2097034478
#>  $ : int [1:7] 10407 -1092868394 2068048389 -941047964 1985001490 281081429 1348762102
#>  $ : int [1:7] 10407 1596893082 662198731 13755946 -476592091 -1148015801 -279127479
#>  $ : int [1:7] 10407 1534427354 1787611233 -2138324094 984649385 -1567589143 1264754606
#>  $ : int [1:7] 10407 1679279718 -1378591168 1075751467 -1383225463 -1579733463 -733897114
#>  $ : int [1:7] 10407 902189764 1930969057 1121338310 722422767 909572148 295945097
#>  $ : int [1:7] 10407 1439705525 -541495023 -680310854 -1338197738 738580423 310732508
#>  $ : int [1:7] 10407 1236898335 -1488386113 379253309 592132010 -815147180 -1841856331
#>  $ : int [1:7] 10407 38994813 542671344 -1152371290 234275939 -541633803 -735359153
#>  $ : int [1:7] 10407 -1822911063 1172332443 1110736837 -817408047 1565328922 -323948131
#>  $ : int [1:7] 10407 -889847563 91177352 441085019 -1287899698 -819401202 1457541669
#>  $ : int [1:7] 10407 1630540097 -509930889 598931520 -346650446 -335132629 -1344326689
#>  $ : int [1:7] 10407 188307014 1775920052 490275801 1139718289 -92597559 1432304691
#>  $ : int [1:7] 10407 -1483464405 1644386517 877331144 1603325223 -1104965090 1959223573
#>  $ : int [1:7] 10407 497249819 -942576417 -2113855584 1889476465 759606460 -1099887295
#>  $ : int [1:7] 10407 1455072933 -2037074541 893849703 -1866660534 -239700167 477798854
#>  $ : int [1:7] 10407 1932150120 -1528776593 -166080570 1616682095 1105158474 -1758706487
#>  $ : int [1:7] 10407 1617803669 -1672126260 1853438443 501777059 -2024820619 -63162295
#>  $ : int [1:7] 10407 -1854517729 1581822533 -1367486781 -481871836 831684560 1509473854
#>  $ : int [1:7] 10407 1170998546 -1197315846 1896830619 893545003 1875528696 -2080089284
#>  $ : int [1:7] 10407 1493236010 -1893148130 1831271712 -506769988 1716829630 981740274
#>  $ : int [1:7] 10407 1882870378 -176591861 1890232275 981456578 2079500075 571601533
#>  $ : int [1:7] 10407 -936879989 -925591853 -494372316 -1673859496 1010849299 -1930015749
#>  $ : int [1:7] 10407 -712524684 686788664 -2069726797 -114626507 -1891074929 -1246968015
#>  $ : int [1:7] 10407 505994383 -72324296 -361258297 602860828 644666311 1830949561
#>  $ : int [1:7] 10407 -398312440 2016625008 337404590 -634482601 -1573998130 741124800
#>  $ : int [1:7] 10407 -1952831831 -1576646675 -220021997 1119875275 -1774815631 -1019203837
#>  $ : int [1:7] 10407 -390150912 -218928331 2136499887 1859503484 -1751625418 330268042
#>  $ : int [1:7] 10407 1184124085 296749493 1871351069 -1532005216 -1229860034 312491272
#>  $ : int [1:7] 10407 1807451960 1697677176 -1593573777 -1797310239 -1574664506 1567901872
#>  $ : int [1:7] 10407 1281946343 1714966683 -1242603563 -1743819556 -1326355930 830749927
#>  $ : int [1:7] 10407 890344678 1369788783 724163849 -238663838 1094721979 -648371283
#>  $ : int [1:7] 10407 -1703221641 -1493313205 483021508 415717044 -1902655806 1080122199
#>  $ : int [1:7] 10407 -915497308 -2059515223 -1776097448 1034798536 -2040301532 -989270196
#>  $ : int [1:7] 10407 -778376855 -815103695 -1396697823 1852989242 -980354355 193753846
#>  $ : int [1:7] 10407 764747096 -109867633 880300473 951894538 -51408312 -1211005396
#>  $ : int [1:7] 10407 -297554105 1994783002 -1073267348 -2083951987 -375320291 -1250369429
#>  $ : int [1:7] 10407 497681901 2078746373 -1160322294 1114988993 -400859323 2121046407
#>  $ : int [1:7] 10407 2115198156 1711229004 1515690079 -300773292 -1597065369 -172166385
#>  $ : int [1:7] 10407 450649395 1550622094 1829128708 2050878606 -1617933634 597522092
#>  $ : int [1:7] 10407 931767521 302041940 247547521 530774089 548399322 -219084080
#>  $ : int [1:7] 10407 712967593 177247368 2037247281 1830389900 1574529688 -1315187761
#>  $ : int [1:7] 10407 2142627432 -1865077488 -60171811 -858137845 -638158549 -515047053
#>  $ : int [1:7] 10407 -1813816063 1749434134 -1472124739 958932093 2108483717 1303570638
#>  $ : int [1:7] 10407 1207360035 1972460077 528278046 1585415864 550672084 -1459382240
#>  $ : int [1:7] 10407 -1808873857 -734941989 910581400 -163624577 -50496662 -1594859829
#>  $ : int [1:7] 10407 325919583 857960318 69609938 -1278826932 583602471 146947970
#>  $ : int [1:7] 10407 -1877382894 -1466795449 -802145263 -970859313 -1229214581 795515025
#>  $ : int [1:7] 10407 -561445567 -2117614224 -2014138489 1464653226 521700729 -1780882196
#>  $ : int [1:7] 10407 1404709348 1625327788 1256442840 -1806248372 777261380 124794048
#>  $ : int [1:7] 10407 -2043052842 500345755 -638992838 -556994382 906476928 -746105445
#>  $ : int [1:7] 10407 1891329238 -327491765 1952179364 -158098131 -229300912 -1152627180
#>  $ : int [1:7] 10407 -1267662236 -299382690 -1242929100 -2112768711 -1650258957 -1975493262
#>  $ : int [1:7] 10407 831237382 216650206 1253992838 -794700063 -538336740 1332472442
#>  $ : int [1:7] 10407 259363718 -1616005318 2066266198 -61818839 -887335199 -463549808
#>  $ : int [1:7] 10407 866157390 1155085760 1216586011 -1570810711 -286066271 909240215
#>  $ : int [1:7] 10407 1402649803 1110137970 656345449 -933645764 -1452845248 -1180179693
#>  $ : int [1:7] 10407 556819550 1191501191 453480173 1446866682 223787454 714493780
#>  $ : int [1:7] 10407 -337959752 -2107281051 1653313460 530627258 -1590967611 695263503
#>  $ : int [1:7] 10407 -1158687954 -1286365641 1779332387 1414514930 277032345 1527591530
#>  $ : int [1:7] 10407 568491903 1148364050 2047386945 1493069188 323877589 108296158
#>  $ : int [1:7] 10407 -1165680509 177185287 -766288976 -717431959 -1065507864 1756254813
#>  $ : int [1:7] 10407 1837135497 1208794314 -114721407 -236829195 -98457053 -14941669
#>  $ : int [1:7] 10407 -1656809956 1385363742 -1029212850 -1075045046 -250206909 -933070103
#>  $ : int [1:7] 10407 -353596452 -963431742 -469937662 631866902 727807158 1449541614
#>   [list output truncated]